Add mem_trace state in kernel-proper
Previously it was possible to switch between different
configurations (fastswap+ssdopt), sync/async writes,
tape/no_tape by reloading the cli with appropriate argument
However, since all of this state was kept in variables
owned by the module, all setting changes needed to be combined
in a single module reload which made the api clunky.

This adds a struct in kernel (in mm/injections.c) which
holds state for mem_trace functionality and can be modified
by injector module. Since this struct is in kernel, it
survives module reloads so individual knobs can be changed
in separate module reloads