Change tape fetch sync strategy to use schedule
Previously the recorded tape was used directly for prefetching
If a process registered to be prefetched, the kernel would load
its tape and maintain 2 pointers: one indicating current position
of the appliction on tape, and another indicating the next
position in which  prefetching batch is to be performed.

This worked well for mmap_random_rw since page addresses were
largely unique and so at least a minor page fault was guaranteed
on every page access.
This, however is not the case in other more realistic workloads.
This is particularly far from truth for mmult_eigen matrix
multiplication workload. Since Eigen is optimized for memory
locality, most memory accesses are repeating in the workload. But
since in this case we only get a page fault on the first access,
the position pointer maintained by the kernel to know where
the application is, quickly falls out of sync.

In the new strategy we preprocess the tape with lru simulator
and create a much smaller tape containing the list of pages
that resulted in a fault according to the lru simulator.

TO start off, we are guaranteed to fault on the first
address in the tape which is used to prefetch the first batch.
After the first and all the other prefetching batches
the kernel chosses the next address which will be used
to respawn prefetcher. This next address is usually the address
after the last prefetched one. Here, in the new approach
the kernel does an additional check to make sure the
address (page fault on which we are waiting to spwan next
batch of prefetching) is not already PRESENT in process
page table as in that case we would not get a fault. If
the chosen address is already present(unlikely), the kernel
walks down the tape until it finds an address not present

Starting and prefetching steps are identical with previous
approach but unlike the previous approach, the kernel no
longer needs to maintain tape state on every page fault